05-21 토큰에러 수정, 게시판 좋아요 구현, 게시판 날짜 및 시간 추가, 진료과목 2개이상 검색 구현 commit

+++ src/main/java/hannyanggang/catchdoctor/repository/CommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    List<Comment> findAllByBoardId(int boardId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findAllByBoardId(Long boardId);
 }
