+++ src/main/java/hannyanggang/catchdoctor/service/CommentService.java
@@ -24,7 +24,7 @@ public class CommentService {
 
     // 댓글 작성하기
     @Transactional
-    public CommentDto writeComment(int boardId, CommentDto commentDto, User user) {
+    public CommentDto writeComment(Long boardId, CommentDto commentDto, User user) {
         Comment comment = new Comment();
         comment.setContent(commentDto.getContent());
 
@@ -44,7 +44,7 @@ public class CommentService {
 
     // 댓글 불러오기 (1개)
     @Transactional(readOnly = true)
-    public CommentDto getComment(int commentId) {
+    public CommentDto getComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> {
             return new IllegalArgumentException("댓글이 존재하지 않습니다.");
         });
@@ -56,7 +56,7 @@ public class CommentService {
 
     // 글에 해당하는 전체 댓글 불러오기
     @Transactional(readOnly = true)
-    public List<CommentDto> getComments(int boardId) {
+    public List<CommentDto> getComments(Long boardId) {
         List<Comment> comments = commentRepository.findAllByBoardId(boardId);
         List<CommentDto> commentDtos = new ArrayList<>();
 
@@ -67,7 +67,7 @@ public class CommentService {
 
     // 댓글 삭제하기
     @Transactional
-    public String deleteComment(int commentId) {
+    public String deleteComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(()-> {
             return new IllegalArgumentException("댓글 Id를 찾을 수 없습니다.");
         });
